+++ lib/moxl/src/Stanza/Muc.php
@@ -10,10 +10,23 @@ class Muc
 {
     public static function message($to, $content = false, $html = false, $id = false,
         $file = false, $parentId = false, array $reactions = [], $originId = false,
-        $threadId = false, $parentThreadId = false)
+        $threadId = false, $mucReceipts = false)
     {
-        Message::maker($to, $content, $html, 'groupchat', false, false, $id, false, $file,
-            false, $parentId, $reactions, $originId, $threadId, $parentThreadId);
+        Message::maker(
+            $to,
+            $content,
+            $html,
+            'groupchat',
+            $mucReceipts, // chatstates required as well
+            $mucReceipts ? 'request' : false,
+            $id,
+            false,
+            $file,
+            false,
+            $parentId,
+            $reactions,
+            $originId,
+            $threadId);
     }
 
     public static function active($to)