+++ lib/moxl/src/Xec/Payload/Displayed.php
@@ -8,10 +8,14 @@ class Displayed extends Payload
     {
         $message = \App\User::me()->messages()
                                   ->where('replaceid', (string)$stanza->attributes()->id)
-                                  ->where('jidfrom', current(explode('/', (string)$parent->attributes()->to)))
+                                  ->where('jidfrom', current(explode('/',
+                                    'groupchat' == (string)$parent->attributes()->type
+                                        ? (string)$parent->attributes()->from
+                                        : (string)$parent->attributes()->to
+                                  )))
                                   ->first();
 
-        if ($message) {
+        if ($message && $message->displayed == null) {
             $message->displayed = gmdate('Y-m-d H:i:s');
             $message->save();